This is a test

+++ Smart shopping system/src/SalesRecord.java
@@ -1,37 +1,36 @@
 package smartshop;
-// Import the Product class from the same package
-import smartshop.Product;  // Add this line to import the Product class
 
-// This class represents a sales record.
 public class SalesRecord {
     private Product product;  // The product that was sold
     private String date;      // The date of the sale
     private int quantity;     // The quantity of the product sold
 
-    // Constructor to initialize sales record
+    // Constructor: Initializes the sales record with product, date, and quantity sold
     public SalesRecord(Product product, String date, int quantity) {
-        this.product = product;
-        this.date = date;
-        this.quantity = quantity;
+        this.product = product;  // Set the product sold
+        this.date = date;        // Set the sale date
+        this.quantity = quantity;  // Set the quantity sold
     }
 
-    // Getters for the sales record details
+    // Getter methods to access the sales record details
     public Product getProduct() {
-        return product;
+        return product;  // Returns the product that was sold
     }
 
     public String getDate() {
-        return date;
+        return date;  // Returns the date of the sale
     }
 
     public int getQuantity() {
-        return quantity;
+        return quantity;  // Returns the quantity of the product sold
     }
 
-    public double getTotalPrice() {
-        return product.getPrice() * quantity;  // Calculate the total price of the sale
+    // This method calculates the total price of the sale (product price * quantity sold)
+    public float getTotalPrice() {
+        return product.getPrice() * quantity;  // Multiply the product price by the quantity sold
     }
 }
+
